refactor: rename layout
