+++ app/javascript/dashboard/routes/dashboard/start/components/OnboardingLayout.vue
